Some Refactoring:
   * move metadata_mvn and metadat_npm modules to maven and npm
   * add one missing requirement
   * use session to create s3 resource
   * add some flag and args for upload and delete command
