+++ requirements-dev.in
@@ -1,3 +1,4 @@
+click
 jinja2
 boto3
 botocore
